+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [15.0.2] - 2023-02-06
+### Added
+- winter_django.create_django_urls_from_routes: function that returns a list of django urls by the list of Routes 
+
 ## [15.0.0] - 2023-02-01
 ### Changed
 - winter.messaging: EventBus interface segregated and renamed to EventPublisher